+++ src/main/kotlin/de/moritzf/opencodewebpanel/settings/OpenCodeSettingsListener.kt
@@ -22,6 +22,8 @@ interface OpenCodeSettingsListener {
 
     fun uiSettingChanged(setting: OpenCodeUiSetting, enabled: Boolean) {}
 
+    fun systemNotificationsChanged(enabled: Boolean) {}
+
     fun serverRestartRequested() {}
 
     companion object {